cleanup

+++ cute_kernels/kernels/rnn/triton_implementation/forward.py
@@ -48,9 +48,9 @@ def rnn_forward_triton_kernel(
 
     for _ in range(S):
         input_ptrs = input_ptr + indices
-        input = tl.load(input_ptrs, mask=mask_bh, other=0).to(tl.float32)
+        input = tl.load(input_ptrs, mask=mask_bh, other=0)
 
-        input_state = tl.dot(input_state, weight, input, allow_tf32=True).to(input_state.dtype)
+        input_state = tl.dot(input_state, weight, input, allow_tf32=True, out_dtype=input_state.dtype)
         input_state = tanh(input_state)
 
         output_ptrs = output_ptr + indices
